major json work
The relatively new "nvme-generic" devices (e.g. /dev/ng0n1)
now processed and shown when the --generic option is given.
With NVMe device single line output, reduce model string
length to 30 (was 41) to align columns better.



git-svn-id: https://svn.bingwo.ca/repos/lsscsi/trunk@177 aaa919c1-9816-4d37-901d-0171dbb97e96

+++ src/sg_pr2serr.c
@@ -118,7 +118,7 @@ sgj_parse_opts(sgj_state * jsp, const char * j_optarg)
     int k, c;
 
     for (k = 0; j_optarg[k]; ++k) {     /* step over leading whitespace */
-        if (! isspace(j_optarg[k]))
+        if (! isspace((uint8_t)j_optarg[k]))
             break;
     }
     for ( ; j_optarg[k]; ++k) {
